% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Get the inferred residual covariance matrix}
\usage{
get_sigma(m)
}
\arguments{
\item{m}{a model fitted with \code{jtdm_fit}}
}
\value{
A list containing:\tabular{ll}{
   \code{Ssamples} \tab Sample from the posterior distribution of the residual covariance matrix. It is an array where the first two dimensions are the rows and columns of the matrix, and the third dimensions are the samples from the posterion distribution \cr
   \tab \cr
   \code{Smean} \tab Posterior mean of the residual covariance matrix. \cr
   \tab \cr
   \code{Sq975,Sq025} \tab 97.5\% and 0.25\% posterior quantiles of the residual covariance matrix. \cr
}
}
\description{
Get the samples from the posterior distribution of the residual covariance matrix, together with the posterior mean and quantiles.
}
\examples{
data(Y)  \cr
data(X)  \cr
# Short MCMC to obtain a fast example: results are unreliable !
m = jtdm_fit(Y=Y, X=X, formula=as.formula("~GDD+FDD+forest"),  adapt = 10,  \cr
        burnin = 100,  \cr
        sample = 100)  \cr
# get the inferred residual covariance
Sigma =get_sigma(m)
}
