% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtdm-package.R
\docType{package}
\name{jtdm-package}
\alias{jtdm}
\alias{jtdm-package}
\title{jtdm.}
\description{
Package to fit a Join Trait Distribution Model and to analyse its result to understand and predict the community-level strategy. See Poggiato et al. In prep.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giopogg/jtdm}
  \item \url{https://giopogg.github.io/jtdm/}
  \item Report bugs at \url{https://github.com/giopogg/jtdm/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Poggiato \email{giov.poggiato@gmail.com} (\href{https://orcid.org/0000-0003-1957-9764}{ORCID}) [copyright holder]

}
\keyword{internal}
