% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtdm.R
\docType{package}
\name{jtdm}
\alias{jtdm}
\title{jtdm.}
\description{
Package to fit a Join Trait Distribution Model and to analyse its result to understand and predict the community-level strategy. See Poggiato et al. In prep.
}
