% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtdmCV.R
\name{jtdmCV}
\alias{jtdmCV}
\title{K-fold cross validation predictions and goodness of fit metrics}
\usage{
jtdmCV(
  m,
  K = 5,
  adapt = 200,
  burnin = 500,
  sample = 500,
  n.chains = 2,
  partition = NULL
)
}
\arguments{
\item{m}{a model fitted with \code{jtdm_fit}}

\item{K}{The number of folds of the K-fold cross validation}

\item{adapt, burnin, sample, n.chains}{Parameters of the MCMC sampler. See \code{?run.jags} for details}

\item{partition}{A partition of the dataset specified by the user. It is a vector (whose length are the number of sites), where each element specifies the fold index of the site.}
}
\value{
A list containing:
   \item{Pred}{Sample from the posterior predictive distribution in cross validation. It is an array where the first dimension is the number of sites in Xnew, the second is the number of traits modelled and the third the number of MCMC samples. NULL if FullPost=FALSE. }
   
   \item{PredMean}{Posterior mean of posterior predictive distribution in cross validation. }
   
   \item{Predq975,Predq025}{97.5\% and 0.25\% posterior quantiles of the posterior predictive distribution in cross validation. NULL if FullPost=FALSE. }
   
   \item{R2}{R squared of predictions in cross validation. }
   
   \item{RMSE}{Root square mean error between  squared of predictions in cross validation.}
}
\description{
Run K-fold cross validation predictions of the model m on a specified dataset.
}
\examples{
data(Y)  
data(X)  
# Short MCMC to obtain a fast example: results are unreliable !
m = jtdm_fit(Y=Y, X=X, formula=as.formula("~GDD+FDD+forest"),  adapt = 1,  
        burnin = 10,  
        sample = 10)  
# Run 3-fold cross validation on m
pred = jtdmCV(m, K=3)
}
