% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtdm.R
\docType{data}
\name{X}
\alias{X}
\title{Site x environmental covariates dataset}
\format{
A matrix
}
\usage{
data(X)

data(X)
}
\description{
Includes the Growing Degree Days (GDD) during the growing season and Freezing Degree Days (FDD) during the growing season averaged over the period 1989-2019
}
\examples{
data(X)
}
\author{
Orchamp consortium
}
\keyword{datasets}
