% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtdm-package.R
\docType{data}
\name{Y}
\alias{Y}
\title{Site x CWM traits dataset}
\format{
A matrix
}
\usage{
data(Y)
}
\description{
A site x CWM traits dataset computed using pinpoint abundances of plants and species mean
}
\examples{
data(Y)
}
\author{
Orchamp Consortium
}
\keyword{datasets}
