% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Get the inferred residual covariance matrix}
\usage{
get_sigma(m)
}
\arguments{
\item{m}{a model fitted with \code{jtdm_fit}}
}
\value{

}
\description{
Get the samples from the posterior distribution of the residual covariance matrix, together with the posterior mean and quantiles.
}
\examples{
data(Y)  
data(X) 
# Short MCMC to obtain a fast example: results are unreliable !
m = jtdm_fit(Y=Y, X=X, formula=as.formula("~GDD+FDD+forest"),  adapt = 10, 
        burnin = 100, 
        sample = 100) 
# get the inferred residual covariance
Sigma =get_sigma(m)
}
